requirements.txt
yt-dlp>=2024.10.0
faster-whisper>=1.1.0
ollama>=0.3.0
python-dotenv>=1.0.0
